Implemented "Waiting..." status for sign-in events
Navigation bar will now properly display "Waiting..." while user is being authenticated. The app must have an initial state of "waiting: true" due to the re-load of the page on google auth sign-in by redirect.

TODO: The app reads state to get current state, which violates React's recommended practice, but I think this is ok because the read is triggered by a callback within Firebase's own "onAuthStateChanged" listener function. But consider changing the waiting setState() as a function according to recommendation